Making, Importing And Using A C Library In Python (Linux Version)
ow to create and import a custom C library in Python (Linux). 

This is something I picked up recently. I am unsure what the best practices are-- but the method I picked up appears to work. What you will need is a compiler (we will be using GCC), and of course, Python.